Crate bmw_deps

The BMW Dependency crate

This is the dependency crate. All BMW dependencies are included in this crate as re-exports and used by the other crates within the repo.

Re-exports

pub use libc;

pub use nix;

pub use backtrace;

pub use base64;

pub use bitvec;

pub use byteorder;

pub use chrono;

pub use colored;

pub use convert_case;

pub use dirs;

pub use downcast;

pub use dyn_clone;

pub use errno;

pub use failure;

pub use failure_derive;

pub use flate2;

pub use futures;

pub use itertools;

pub use lazy_static;

pub use math;

pub use num_format;

pub use path_clean;

pub use portpicker;

pub use rand;

pub use rand_core;

pub use random_string;

pub use ring;

pub use sha1;

pub use substring;

pub use url;

pub use url_path;